Add Ultra Cosmetics plugin!
This will be a great plugin for players on the MineHut server and would recommend a player having Tier 1+ for this because of such cool features (Not making fun of poor people) and this could lead to more people buying tiers and more money for you guys. Hope you add it! And I just want to tell you that I'm unsure of the plugin config name.

More Info: https://www.spigotmc.org/resources/ultra-cosmetics-opensource-and-free-1-8-1-11.10905/